+++ CHANGES.rst
@@ -3,6 +3,8 @@
 
 Updates
 .......
+- LineList.available_transitions() no longer has key argument n_max
+- LineList: extra attributes for transitions added (`ion_name`, `log(w*f)`, `abundance`, `ion_correction`, `rel_strength`)
 - ASCII tables with no header are required to be 4 columns or less for `io.readspec` to work
 - Modify XSpectrum1D to use masked numpy arrays
 - Enable hdf5 I/O  [requires h5py]